ask-my-docs-requirements.txt
streamlit
python-dotenv
langchain
langchain-classic
langchain-google-genai
langchain-community
langchain-text-splitters
langchain-core
faiss-cpu
pypdf
